backend_server/requirements.txt
jsonrpclib-pelix
tap_news_utils
redis
